pyproject.toml
[tool.poetry]
name = "shell-daemon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
tts = ">=0.22.0,<0.23.0"
llama-cpp-python = ">=0.3.12,<0.4.0"
torch = ">=2.7.1,<3.0.0"
torchaudio = ">=2.7.1,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
